+++ src/DAO/DAO.java
@@ -27,7 +27,7 @@ public abstract class DAO<T> {
      * @param id
      * @return
      */
-    public abstract T find(long id);
+    public abstract T find(int id);
 
     /**
      * Permet de créer une entrée dans la base de données par rapport à un objet
