+++ src/DAO/MaladeDAO.java
@@ -19,7 +19,7 @@ public class MaladeDAO extends DAO<Malade>{
     }
 
     @Override
-    public Malade find(long id) {
+    public Malade find(int id) {
         throw new UnsupportedOperationException("Not supported yet."); //To change body of generated methods, choose Tools | Templates.
     }
