+++ src/DAO/ServiceDAO.java
@@ -19,7 +19,7 @@ public class ServiceDAO extends DAO<Service>{
     }
 
     @Override
-    public Service find(long id) {
+    public Service find(int id) {
   	   throw new UnsupportedOperationException("Not supported yet."); //To change body of generated methods, choose Tools | Templates.
     }
